render_requirements.txt
python-telegram-bot==20.7
cairosvg==2.8.2
Pillow==11.3.0
lottie==0.7.2